+++ src/py/mat3ra/made/tools/utils/factories.py
@@ -1,7 +1,7 @@
 from mat3ra.utils.factory import BaseFactory
 
 
-class PerturbationFunctionHelperFactory(BaseFactory):
+class PerturbationFunctionHolderFactory(BaseFactory):
     __class_registry__ = {
-        "sine_wave": "mat3ra.made.tools.utils.helpers.SineWave",
+        "sine_wave": "mat3ra.made.tools.utils.functions.SineWaveFunctionHolder",
     }
